[feat]: [CDS-79241] Variables auto-suggestions for custom stage environment (#53115)
* [feat]: [CDS-79241] variable creator changes

* [feat]: [CDS-78377] Fixing junit test

* [feat]: [CDS-79241] inject variable creator in plan creator class

* [feat]: [CDS-79241] fix junit

* [feat]: [CDS-79241] remove duplicate code
